Remove Boost dependency (BROKEN! FIX THIS!)
I tried to remove the Boost::regex dependency by switching to std::regex, but
now there is a problem with the AC records in the OpenAir files.  If an airspace
class is defined as follows:

AC C/E

Then it doesn't get parsed as class 'C/E', but it does get parsed as 'C'.  So we
should somehow use the 'longest match' to make this work.  See also the regular
expression for REGEX_AIRSPACE_CLASS in Parser.cpp.